backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.5
astroid==2.3.3
autopep8==1.5
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
distlib==0.3.0
Django==3.0.4
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
filelock==3.0.12
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.2.1
mccabe==0.6.1
multitasking==0.0.9
numpy==1.18.3
pandas==1.0.3
pycodestyle==2.5.0
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
responses==0.10.12
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
virtualenv==20.0.13
wrapt==1.11.2
yfinance==0.1.54
